qcacld-3.0: Correct the tlv length of the hangdata
The Tlv of the hang data is reduced to 2bytes from the 4bytes
change the structure defnitions accordingly.

CRs-Fixed: 2753804

+++ core/mac/src/pe/lim/lim_api.c
@@ -79,7 +79,7 @@
 #include <qdf_notifier.h>
 
 struct pe_hang_event_fixed_param {
-	uint32_t tlv_header;
+	uint16_t tlv_header;
 	uint8_t vdev_id;
 	uint8_t limmlmstate;
 	uint8_t limprevmlmstate;
